# ai-reply-back-end/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Reply Backend API
  version: "1.0.0"
  description: >
    Мобильный API AI Reply. Тексты сообщений не сохраняются: сервер обрабатывает
    их только в памяти во время запроса.
...
servers:
  - url: https://ai-reply.kz
    description: production
paths:
...
  /api/v1/auth/email/otp/request:
    post:
      summary: Отправить 4-значный код на почту (Resend)
      description: >
        Ответ одинаков для известного и нового адреса. Новый код аннулирует
        предыдущий; повторный запрос раньше OTP_RESEND_COOLDOWN отклоняется.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email, example: "aigerim@mail.kz" }
                locale: { type: string, enum: [kk, ru, en, uz] }
      responses:
        "200":
          description: Код отправлен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailChallenge" }
        "400": { description: INVALID_EMAIL }
        "429":
          description: >-
            OTP_RESEND_COOLDOWN (details.retry_after_seconds, Retry-After) / RATE_LIMITED — лимиты на адрес
            (RATE_OTP_REQUEST_PER_ADDRESS_PER_HOUR, RATE_OTP_REQUEST_PER_DAY, OTP_MAX_FAILED_PER_DAY неверных
            кодов за 24 ч; с retry_after_seconds) и мягкий лимит на IP (RATE_OTP_REQUEST_PER_HOUR)
        "503": { description: EMAIL_DELIVERY_FAILED }
  /api/v1/auth/email/otp/verify:
    post:
      summary: Проверить код и получить сессию (аккаунт создаётся при первом входе)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{4}$", example: "0384" }
                device: { $ref: "#/components/schemas/Device" }
      responses:
        "200": { description: "access_token, refresh_token, user (auth_providers), profile, subscription, usage" }
        "400": { description: "INVALID_OTP (details.attempts_remaining) / OTP_EXPIRED / OTP_ALREADY_USED / INVALID_EMAIL" }
        "429": { description: "OTP_ATTEMPTS_EXCEEDED / RATE_LIMITED (OTP_MAX_FAILED_PER_DAY неверных кодов на адрес за 24 ч — details.retry_after_seconds; IP — RATE_OTP_VERIFY_PER_HOUR)" }
  /api/v1/auth/google:
    post:
      summary: Вход через Google (ID token проверяется сервером по JWKS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token, nonce]
              properties:
                id_token: { type: string }
                nonce: { type: string, description: "Тот же nonce, что передан в Google SDK / Credential Manager" }
                device: { $ref: "#/components/schemas/Device" }
      responses:
        "200": { description: Сессия }
        "401": { description: INVALID_ID_TOKEN — не истёкшая сессия, refresh не нужен }
        "409": { description: EMAIL_ALREADY_IN_USE }
        "503": { description: AUTH_PROVIDER_UNAVAILABLE }
  /api/v1/auth/apple:
    post:
      summary: Вход через Apple (identity token проверяется сервером по JWKS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identity_token, nonce]
              properties:
                identity_token: { type: string }
                nonce: { type: string, description: "Сырой nonce; Apple получил его SHA-256 (hex)" }
                full_name: { type: string, maxLength: 120 }
                device: { $ref: "#/components/schemas/Device" }
      responses:
        "200": { description: Сессия }
        "401": { description: INVALID_ID_TOKEN }
        "409": { description: EMAIL_ALREADY_IN_USE }
        "503": { description: AUTH_PROVIDER_UNAVAILABLE }
  /api/v1/auth/request-otp:
    post:
      summary: "Устаревший вход по identifier (старые сборки)"
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier: { type: string, example: "+7 701 123 45 67" }
                locale: { type: string, enum: [kk, ru, en, uz] }
      responses:
        "200": { description: Запрос принят, код отправлен доступным каналом }
        "400": { description: INVALID_REQUEST — номер не поддерживается }
        "429": { description: RATE_LIMITED }
  /api/v1/auth/verify-otp:
    post:
      summary: "Устаревшая проверка кода по identifier (старые сборки)"
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, code]
              properties:
                identifier: { type: string }
                code: { type: string, example: "4827" }
                device:
                  type: object
                  properties:
                    device_id: { type: string }
                    platform: { type: string, enum: [ios, android, web] }
                    app_version: { type: string }
                    os_version: { type: string }
                    locale: { type: string }
                    timezone: { type: string }
      responses:
        "200": { description: access_token, refresh_token, user, usage }
        "400": { description: INVALID_OTP / OTP_EXPIRED }
  /api/v1/auth/refresh:
    post:
      summary: Обновить пару токенов (с ротацией)
      responses:
        "200": { description: OK }
        "401": { description: UNAUTHORIZED }
  /api/v1/auth/logout:
    post:
      summary: Закрыть сессию (вся цепочка refresh-токена, в т.ч. уже ротированного)
      parameters:
        - name: X-Installation-ID
          in: header
          required: false
          description: >-
            Установка, с которой выходят: она сразу отвязывается от аккаунта (даже если
            refresh-токен уже истёк). Единственный запрос, где клиент шлёт этот заголовок.
          schema: { type: string, pattern: "^[A-Za-z0-9-]{8,64}$" }
      responses: { "200": { description: OK } }
  /api/v1/me:
    get:
      summary: Профиль, тариф и квота (user.preferred_language — kk | ru | en | uz | "")
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
    delete:
      summary: Удалить аккаунт и все его данные на сервере (сразу, без отмены)
      description: >-
        Тело необязательно. С apple_authorization_code сервер отзывает токен Sign in with Apple,
        если у аккаунта есть вход через Apple и заданы APPLE_TEAM_ID / APPLE_KEY_ID / APPLE_PRIVATE_KEY;
        ошибка Apple удалению не мешает. После ответа все токены аккаунта отвечают 401.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteAccountRequest" }
      responses:
        "200":
          description: Удалено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteAccountResult" }
        "400": { description: INVALID_REQUEST — неизвестное поле }
        "401": { description: UNAUTHORIZED }
    patch:
      summary: Частичное обновление профиля (POST /api/v1/me — то же для Android)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { description: "INVALID_REQUEST — неизвестное поле, неверный grammatical_gender, onboarding_version вне 0…1000 или неизвестный preferred_language (details.field)" }
  /api/v1/me/delete:
    post:
      summary: То же, что DELETE /api/v1/me (для клиентов, которые не шлют тело с DELETE)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteAccountRequest" }
      responses:
        "200":
          description: Удалено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteAccountResult" }
        "401": { description: UNAUTHORIZED }
  /api/v1/account/delete/request:
    post:
      summary: Страница /account/delete — код удаления на почту аккаунта
      description: >-
        Ответ всегда {"ok": true}: существует ли аккаунт, не раскрывается. Код (purpose=delete)
        уходит только на почту существующего аккаунта, с лимитами и паузой обычного e-mail OTP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              additionalProperties: false
              properties:
                email: { type: string, format: email }
                locale: { type: string, enum: [kk, ru, en, uz] }
      responses:
        "200": { description: "{ ok: true }" }
        "400": { description: INVALID_EMAIL }
        "429": { description: RATE_LIMITED }
  /api/v1/account/delete/confirm:
    post:
      summary: Страница /account/delete — код верен, аккаунт удаляется
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              additionalProperties: false
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{4}$" }
      responses:
        "200": { description: "{ deleted: true }" }
        "400": { description: "INVALID_OTP (details.attempts_remaining) / OTP_EXPIRED / OTP_ALREADY_USED" }
        "429": { description: OTP_ATTEMPTS_EXCEEDED / RATE_LIMITED }
  /api/v1/me/email/otp/request:
    post:
      summary: Код на почту, которую нужно добавить к аккаунту для входа
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                locale: { type: string, enum: [kk, ru, en, uz] }
      responses:
        "200":
          description: Код отправлен
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailChallenge" }
        "409": { description: CONFLICT — у аккаунта уже есть почта / EMAIL_ALREADY_IN_USE }
        "429": { description: OTP_RESEND_COOLDOWN / RATE_LIMITED }
  /api/v1/me/email/otp/verify:
    post:
      summary: Подтвердить код — почта привязана к аккаунту
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, pattern: "^[0-9]{4}$" }
      responses:
        "200": { description: То же, что GET /api/v1/me }
        "400": { description: INVALID_OTP / OTP_EXPIRED / OTP_ALREADY_USED }
        "409": { description: EMAIL_ALREADY_IN_USE }
  /api/v1/me/usage:
    get:
      summary: Остаток генераций
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/me/subscription:
    get:
      summary: Текущая подписка
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/me/consents:
    post:
      summary: Сохранить принятие текущей оферты и политики конфиденциальности
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LegalConsentRequest" }
      responses:
        "200":
          description: Согласие сохранено
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LegalConsent" }
        "400": { description: INVALID_REQUEST — версия не текущая (клиент должен обновить конфигурацию) }
    delete:
      summary: Отозвать согласие на AI-обработку
      description: >-
        Идемпотентно. Запись остаётся с withdrawn_at; /ai/reply, /ai/compose и /ai/polish отвечают
        403 CONSENT_REQUIRED, пока приложение снова не отправит POST. Аккаунт сохраняется.
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: "{ ok: true }" }
        "401": { description: UNAUTHORIZED }
  /api/v1/devices:
    post:
      summary: Регистрация устройства (старые сборки; push — через /api/v1/installations)
      deprecated: true
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: OK } }
  /api/v1/installations:
    post:
      summary: Регистрация установки приложения (анонимной или с аккаунтом) и FCM-токена
      description: >-
        features.installations. Аккаунт берётся только из Bearer-токена: с токеном установка
        привязывается к аккаунту, без него становится анонимной; недействительный токен → 401,
        ничего не меняется. Токен хранится зашифрованным и в ответах не возвращается. Лимит —
        RATE_GENERIC_PER_MINUTE на установку и ×20 на IP.
      security: [{ bearerAuth: [] }, {}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InstallationRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Installation" }
        "400": { description: "INVALID_REQUEST (details.field: installation_id, platform, push.provider, push.token, …) или неизвестное поле" }
        "401": { description: недействительный access-токен }
        "429": { description: RATE_LIMITED }
  /api/v1/installations/{installation_id}/detach:
    post:
      summary: Отвязать установку от аккаунта (с токеном — только свою)
      security: [{ bearerAuth: [] }, {}]
      parameters:
        - { name: installation_id, in: path, required: true, schema: { type: string, pattern: "^[A-Za-z0-9-]{8,64}$" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detached: { type: boolean }
        "400": { description: "INVALID_REQUEST (details.field = installation_id)" }
  /api/v1/me/notification-preferences:
    get:
      summary: Категории уведомлений пользователя
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
    put:
      summary: Включить или выключить категории (push и письма, кроме письма subscription_activated; security выключить нельзя)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preferences]
              additionalProperties: false
              properties:
                preferences:
                  type: object
                  minProperties: 1
                  additionalProperties: { type: boolean }
                  example: { marketing: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
        "400": { description: "INVALID_REQUEST — пустой объект, неизвестная категория или security: false (details.field)" }
  /api/v1/notifications/opened:
    post:
      summary: Пользователь открыл push (по did из data)
      description: >-
        Отмечается, только если доставка ушла на эту установку. Вызов best-effort, без повторов.
        Лимит — общий с /installations бюджет установки (по installation_id) и ×20 на IP.
      security: [{ bearerAuth: [] }, {}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [installation_id, delivery_id]
              additionalProperties: false
              properties:
                installation_id: { type: string, pattern: "^[A-Za-z0-9-]{8,64}$" }
                delivery_id: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  recorded: { type: boolean }
        "400": { description: INVALID_REQUEST (details.field) }
  /api/v1/devices/{id}:
    delete:
      summary: Отозвать устройство
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: OK } }
  /api/v1/ai/reply:
    post:
      summary: Сгенерировать ответ
      description: >-
        Язык ответа решает входящее сообщение (source_text), а не язык интерфейса,
        раскладка или язык instruction: profile.reply_language → уверенно определённый
        язык сообщения → input_language → language. Другой язык даёт только явная
        просьба в instruction («ответь на русском»).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_text]
              properties:
                source_text: { type: string, maxLength: 400, description: "max_source_characters из /config" }
                instruction: { type: string, maxLength: 400 }
                language: { type: string, enum: [kk, ru, en, uz], description: "язык интерфейса; язык ответа решает, только если язык сообщения не ясен" }
                input_language: { type: string, enum: [kk, ru, en], description: "раскладка при нажатии Reply (features.sender_profile); решает, только если язык сообщения не ясен; неизвестное значение игнорируется" }
                template_id: { type: string }
                template: { type: object }
                business_context: { type: object }
                profile:
                  type: object
                  properties:
                    description: { type: string }
                    role: { type: string }
                    preferred_tone: { type: string, enum: [natural, friendly, professional, formal, short] }
                    business: { type: object }
                    reply_language: { type: string, enum: [auto, kk, ru, en, uz], description: "features.reply_preferences" }
                    grammatical_gender: { type: string, enum: [male, female, unspecified], description: "features.sender_profile; иначе — сохранённое значение" }
                platform: { type: string }
                app_version: { type: string }
      responses:
        "200":
          description: reply + usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
                  detected_language: { type: string }
                  usage: { $ref: "#/components/schemas/Usage" }
        "400": { description: "INVALID_REQUEST (details.field = source_text) или неизвестное поле" }
        "403": { description: CONSENT_REQUIRED — нет согласия на текущие версии условий и политики }
        "429": { description: "DAILY_LIMIT_REACHED / MONTHLY_LIMIT_REACHED (details: daily_limit, used_today, resets_at; monthly_limit и used_month при месячном лимите) / RATE_LIMITED" }
        "502": { description: AI_PROVIDER_UNAVAILABLE }
        "504": { description: AI_TIMEOUT }
  /api/v1/ai/compose:
    post:
      summary: Написать новое сообщение по инструкции (режим «Создать»)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instruction]
              properties:
                instruction: { type: string, maxLength: 400, description: "max_instruction_length из /config" }
                language: { type: string, enum: [kk, ru, en, uz], description: "язык интерфейса" }
                input_language: { type: string, enum: [kk, ru, en], description: "features.sender_profile" }
                profile:
                  type: object
                  description: "features.sender_profile; других полей нет"
                  properties:
                    grammatical_gender: { type: string, enum: [male, female, unspecified] }
                regenerate: { type: boolean, description: "другая версия по той же инструкции" }
                platform: { type: string }
                app_version: { type: string }
      responses:
        "200":
          description: text + usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  text: { type: string }
                  detected_language: { type: string }
                  usage: { $ref: "#/components/schemas/Usage" }
        "400": { description: "INVALID_REQUEST (details.field = instruction)" }
        "401": { description: UNAUTHORIZED }
        "403": { description: CONSENT_REQUIRED }
        "429": { description: DAILY_LIMIT_REACHED / MONTHLY_LIMIT_REACHED / RATE_LIMITED }
        "502": { description: AI_PROVIDER_UNAVAILABLE }
        "504": { description: AI_TIMEOUT }
  /api/v1/ai/polish:
    post:
      summary: Подсказка — исправленный текст инструкции (квота не тратится)
      description: >-
        Работает, когда features.instruction_polish = true; иначе 404 NOT_FOUND. Потолок —
        LIMIT_POLISH_PER_DAY (100) на пользователя за сутки сервера; сверх него ответ
        200 { text: исходный, changed: false } без обращения к провайдеру.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 400, description: "max_instruction_length из /config" }
                input_language: { type: string, enum: [kk, ru, en] }
                platform: { type: string }
                app_version: { type: string }
      responses:
        "200":
          description: "changed = false — подсказки нет, text равен исходному"
          content:
            application/json:
              schema:
                type: object
                required: [text, changed]
                properties:
                  text: { type: string }
                  changed: { type: boolean }
        "400": { description: "INVALID_REQUEST (details.field = text, details.max_characters)" }
        "401": { description: UNAUTHORIZED }
        "404": { description: NOT_FOUND — polish выключен (AI_POLISH_ENABLED=false) }
        "429": { description: RATE_LIMITED — отдельный лимит RATE_POLISH_PER_MINUTE }
        "502": { description: AI_PROVIDER_UNAVAILABLE }
        "504": { description: AI_TIMEOUT }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminSession:
      type: apiKey
      in: cookie
      name: aireply_admin
      description: Cookie-сессия админ-панели (ADMIN_COOKIE_NAME); для POST нужен ещё заголовок X-CSRF-Token.
  schemas:
...
    ProfileUpdate:
      type: object
      additionalProperties: false
      properties:
        display_name: { type: string }
        role: { type: string }
        description: { type: string }
        preferred_tone: { type: string, enum: [natural, friendly, professional, formal, short] }
        business_offering: { type: string }
        business_summary: { type: string }
        business_rules: { type: array, items: { type: string }, maxItems: 8 }
        locale: { type: string, enum: [kk, ru, en, uz] }
        timezone: { type: string }
        onboarding_completed: { type: boolean }
        grammatical_gender: { type: string, enum: [male, female, unspecified], description: "features.sender_profile" }
        onboarding_version: { type: integer, minimum: 0, maximum: 1000, description: "features.sender_profile; хранится максимум" }
        preferred_language: { type: string, enum: [kk, ru, en, uz], description: "features.preferred_language; ru-KZ → ru, иное → 400. Язык push и писем" }
    Device:
      type: object
      properties:
        device_id: { type: string }
        platform: { type: string, enum: [ios, android, web] }
        app_version: { type: string }
        os_version: { type: string }
        locale: { type: string }
        timezone: { type: string }
...
    DeleteAccountRequest:
      type: object
      additionalProperties: false
      properties:
        apple_authorization_code: { type: string, description: "свежий authorization code Sign in with Apple (iOS)" }
    DeleteAccountResult:
      type: object
      properties:
        deleted: { type: boolean }
        apple_token_revoked: { type: boolean }
...
    EmailChallenge:
      type: object
      required: [masked_email, expires_in, resend_after, code_length]
      properties:
        masked_email: { type: string, example: "a***m@mail.kz" }
        expires_in: { type: integer, example: 300 }
        resend_after: { type: integer, example: 32 }
        code_length: { type: integer, example: 4 }
...