# ai-reply-back-end/internal/transport/api/server.go
package api

import (
	"context"
	"log/slog"
	"net/http"
	"time"

	"github.com/aireply/ai-reply-back-end/config"
	"github.com/aireply/ai-reply-back-end/internal/ai"
	"github.com/aireply/ai-reply-back-end/internal/auth"
	"github.com/aireply/ai-reply-back-end/internal/installations"
	"github.com/aireply/ai-reply-back-end/internal/limits"
	"github.com/aireply/ai-reply-back-end/internal/middleware"
	"github.com/aireply/ai-reply-back-end/internal/notifications"
	"github.com/aireply/ai-reply-back-end/internal/payments"
	"github.com/aireply/ai-reply-back-end/internal/plans"
	"github.com/aireply/ai-reply-back-end/internal/productevents"
	"github.com/aireply/ai-reply-back-end/internal/reports"
	"github.com/aireply/ai-reply-back-end/internal/subscriptions"
	"github.com/aireply/ai-reply-back-end/internal/transport/httpx"
	"github.com/aireply/ai-reply-back-end/internal/users"
)

// Pinger — дерекқордың дайындығын тексеру.
type Pinger func(context.Context) error

// Server — мобильді API.
type Server struct {
	cfg           config.Config
	auth          *auth.Service
	users         *users.Service
	plans         *plans.Service
	subs          *subscriptions.Service
	ai            *ai.Service
	limits        *limits.Service
	payments      *payments.Service
	events        *productevents.Service
	installations *installations.Service
	notifications *notifications.Service
	reports       *reports.Service
	limiter       *middleware.Limiter
	ping          Pinger
	log           *slog.Logger
}

// Deps — сервер тәуелділіктері.
type Deps struct {
	Config        config.Config
	Auth          *auth.Service
	Users         *users.Service
	Plans         *plans.Service
	Subs          *subscriptions.Service
	AI            *ai.Service
	Limits        *limits.Service
	Payments      *payments.Service
	Events        *productevents.Service
	Installations *installations.Service
	Notifications *notifications.Service
	// Reports — AI жауабына шағымдар (POST /api/v1/ai/reports).
	Reports *reports.Service
	Limiter *middleware.Limiter
	Ping    Pinger
	Log     *slog.Logger
}

// New — API сервері.
func New(d Deps) *Server {
	return &Server{
		cfg: d.Config, auth: d.Auth, users: d.Users, plans: d.Plans, subs: d.Subs,
		ai: d.AI, limits: d.Limits, payments: d.Payments, events: d.Events,
		installations: d.Installations, notifications: d.Notifications, reports: d.Reports,
		limiter: d.Limiter, ping: d.Ping, log: d.Log,
	}
}

// Register — маршруттарды негізгі mux-қа қосады.
func (s *Server) Register(mux *http.ServeMux) {
	// Денсаулық тексерулері — балансерге арналған, аутентификациясыз. Орта аты
	// (production, staging) айтылмайды: ол сырттағы ешкімге керек емес.
	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, r *http.Request) {
		httpx.JSON(w, http.StatusOK, map[string]any{"ok": true})
	})
	mux.HandleFunc("GET /readyz", func(w http.ResponseWriter, r *http.Request) {
		if s.ping != nil {
			if err := s.ping(r.Context()); err != nil {
				httpx.JSON(w, http.StatusServiceUnavailable, map[string]any{"ok": false})
				return
			}
		}
		httpx.JSON(w, http.StatusOK, map[string]any{"ok": true})
	})

	ip := func(r *http.Request) string { return httpx.ClientIP(r, s.cfg.App.TrustProxy) }
	limits := s.cfg.Limits

	otpRequest := middleware.RateLimit(s.limiter, "otp_request", limits.OTPRequestPerHour, time.Hour, ip)
	otpVerify := middleware.RateLimit(s.limiter, "otp_verify", limits.OTPVerifyPerHour, time.Hour, ip)
	generic := middleware.RateLimit(s.limiter, "generic", limits.GenericPerMinute, time.Minute, ip)
	perUser := func(r *http.Request) string {
		if u, ok := UserFrom(r.Context()); ok {
			return u.ID
		}
		return httpx.ClientIP(r, s.cfg.App.TrustProxy)
	}
	aiLimit := middleware.RateLimit(s.limiter, "ai", limits.AIPerMinute, time.Minute, perUser)
	// Polish үзіліс сайын шақырылады: жеке шелек, генерация лимитін жемейді.
	polishLimit := middleware.RateLimit(s.limiter, "polish", limits.PolishPerMinute, time.Minute, perUser)
	// Өнім оқиғалары: өз шелегі, басқа сұраныстардың лимитін жемейді.
	eventsLimit := middleware.RateLimit(s.limiter, "events", limits.EventsPerMinute, time.Minute, perUser)
	// Шағымдар: сағатына бір қолданушыға шектеулі (өз шелегі).
	reportsLimit := middleware.RateLimit(s.limiter, "ai_reports", limits.ReportsPerHour, time.Hour, perUser)

	// --- аутентификация: пошта OTP (Resend), Google, Apple
	mux.Handle("POST /api/v1/auth/email/otp/request", otpRequest(http.HandlerFunc(s.handleEmailOTPRequest)))
	mux.Handle("POST /api/v1/auth/email/otp/verify", otpVerify(http.HandlerFunc(s.handleEmailOTPVerify)))
	mux.Handle("POST /api/v1/auth/google", generic(http.HandlerFunc(s.handleGoogleSignIn)))
	mux.Handle("POST /api/v1/auth/apple", generic(http.HandlerFunc(s.handleAppleSignIn)))

	// Бұрынғы идентификатор ағыны: шыққан build-тер (телефон және пошта) үшін сақталған.
	mux.Handle("POST /api/v1/auth/request-otp", otpRequest(http.HandlerFunc(s.handleRequestOTP)))
	mux.Handle("POST /api/v1/auth/verify-otp", otpVerify(http.HandlerFunc(s.handleVerifyOTP)))
	mux.Handle("POST /api/v1/auth/refresh", generic(http.HandlerFunc(s.handleRefresh)))
	mux.Handle("POST /api/v1/auth/logout", generic(http.HandlerFunc(s.handleLogout)))

	// --- профиль және қолданыс
	mux.Handle("GET /api/v1/me", s.requireUser(http.HandlerFunc(s.handleMe)))
	mux.Handle("PATCH /api/v1/me", s.requireUser(http.HandlerFunc(s.handleUpdateMe)))
	// Android alias: HttpURLConnection refuses PATCH outright, and a reflection
	// hack on a platform class is a worse thing to ship than one extra route.
	mux.Handle("POST /api/v1/me", s.requireUser(http.HandlerFunc(s.handleUpdateMe)))
	mux.Handle("GET /api/v1/me/usage", s.requireUser(http.HandlerFunc(s.handleUsage)))
	mux.Handle("GET /api/v1/me/subscription", s.requireUser(http.HandlerFunc(s.handleSubscription)))
	mux.Handle("GET /api/v1/me/devices", s.requireUser(http.HandlerFunc(s.handleListDevices)))
	mux.Handle("POST /api/v1/me/consents", s.requireUser(http.HandlerFunc(s.handleSaveLegalConsent)))
	// Келісімді кері қайтару: AI тоқтайды, тіркелгі қалады.
	mux.Handle("DELETE /api/v1/me/consents", s.requireUser(http.HandlerFunc(s.handleWithdrawLegalConsent)))
	// Тіркелгіні жою. POST — DELETE-ке дене жібере алмайтын клиенттер үшін, мағынасы бірдей.
	mux.Handle("DELETE /api/v1/me", s.requireUser(generic(http.HandlerFunc(s.handleDeleteAccount))))
	mux.Handle("POST /api/v1/me/delete", s.requireUser(generic(http.HandlerFunc(s.handleDeleteAccount))))
	// Қолданбасыз жою (/account/delete беті): поштаға код → код → жою.
	mux.Handle("POST /api/v1/account/delete/request", otpRequest(http.HandlerFunc(s.handleAccountDeletionRequest)))
	mux.Handle("POST /api/v1/account/delete/confirm", otpVerify(http.HandlerFunc(s.handleAccountDeletionConfirm)))
	// Поштасы жоқ тіркелгіге (телефонмен ашылған) пошта қосу — кейін сол поштамен кіру үшін.
	mux.Handle("POST /api/v1/me/email/otp/request", s.requireUser(otpRequest(http.HandlerFunc(s.handleLinkEmailRequest))))
	mux.Handle("POST /api/v1/me/email/otp/verify", s.requireUser(otpVerify(http.HandlerFunc(s.handleLinkEmailVerify))))

	// --- каталог
	mux.Handle("GET /api/v1/plans", generic(http.HandlerFunc(s.handlePlans)))
	mux.Handle("GET /api/v1/config", generic(http.HandlerFunc(s.handleConfig)))

	// --- құрылғылар (ескі build-тер) және орнатулар (push, метадерек)
	mux.Handle("POST /api/v1/devices", s.requireUser(http.HandlerFunc(s.handleRegisterDevice)))
	mux.Handle("DELETE /api/v1/devices/{id}", s.requireUser(http.HandlerFunc(s.handleDeleteDevice)))
	mux.Handle("POST /api/v1/installations",
		s.installationLimit(s.optionalUser(http.HandlerFunc(s.handleRegisterInstallation))))
	mux.Handle("POST /api/v1/installations/{installation_id}/detach",
		s.installationLimit(s.optionalUser(http.HandlerFunc(s.handleDetachInstallation))))
	mux.Handle("GET /api/v1/me/notification-preferences", s.requireUser(http.HandlerFunc(s.handleNotificationPreferences)))
	mux.Handle("PUT /api/v1/me/notification-preferences", s.requireUser(http.HandlerFunc(s.handleUpdateNotificationPreferences)))
	mux.Handle("POST /api/v1/notifications/opened",
		s.installationLimit(s.optionalUser(http.HandlerFunc(s.handleNotificationOpened))))

	// --- AI: мәтін OpenAI-ға тек ағымдағы шарттарға келісім болса кетеді.
	mux.Handle("POST /api/v1/ai/reply", s.requireUser(s.requireConsent(aiLimit(http.HandlerFunc(s.handleReply)))))
	// Нұсқау бойынша жаңа хабарлама («Create»). Сол квота, сол rate limit.
	mux.Handle("POST /api/v1/ai/compose", s.requireUser(s.requireConsent(aiLimit(http.HandlerFunc(s.handleCompose)))))
	// Нұсқауды түзету ұсынысы. Квотасыз; өшірулі болса 404 NOT_FOUND.
	mux.Handle("POST /api/v1/ai/polish", s.requireUser(s.requireConsent(polishLimit(http.HandlerFunc(s.handlePolish)))))
	// AI жауабына шағым. Келісімсіз де қабылданады: шағым OpenAI-ға кетпейді.
	mux.Handle("POST /api/v1/ai/reports", s.requireUser(reportsLimit(http.HandlerFunc(s.handleAIReport))))

	// --- өнім оқиғалары (тек қолданбадан; пернетақта ештеңе жібермейді)
	mux.Handle("POST /api/v1/analytics/events", s.requireUser(eventsLimit(http.HandlerFunc(s.handleProductEvents))))

	// --- төлемдер (демо адаптер)
	mux.Handle("POST /api/v1/payments/checkout", s.requireUser(http.HandlerFunc(s.handleCheckout)))
	mux.Handle("POST /api/v1/payments/{id}/confirm", s.requireUser(http.HandlerFunc(s.handleConfirmPayment)))

	// --- ескі мобильді build-тер (өзгеріссіз жұмыс істейді)
	if s.cfg.Auth.LegacyEnabled {
		mux.Handle("POST /v1/auth/register", otpRequest(http.HandlerFunc(s.handleLegacyRegister)))
		mux.Handle("POST /v1/reply/generate", s.requireLegacy(aiLimit(http.HandlerFunc(s.handleLegacyGenerate))))
	}
}